Make IntegrationMethod scoped enum

+++ kratos/tests/cpp_tests/geometries/test_geometry_data.cpp
@@ -24,7 +24,7 @@ namespace Testing {
 /// Test number of integration methods
 KRATOS_TEST_CASE_IN_SUITE(GeometryDataNumberOfIntegrationMethods, KratosCoreGeometriesFastSuite)
 {
-    const std::size_t n_integration_methods = GeometryData::IntegrationMethod::NumberOfIntegrationMethods;
+    const std::size_t n_integration_methods = static_cast<int>(GeometryData::IntegrationMethod::NumberOfIntegrationMethods);
     // auto static_cast<GeometryData::IntegrationMethod>(10)
 
     KRATOS_CHECK_EQUAL(n_integration_methods, 10);
